requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
redis
python-dotenv
langdetect
edge-tts
google-generativeai